refactor: move Category enum to post package
